refactor volunteer display out of edit_event for reuse with staff
